backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.0
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1